Fix data type for rtu_baud_rate.
Baud rates can exceed the maximum value for uint16_t (e.g. 115200).
Use a uint32_t to fit these larger values.

+++ runtime/core/modbusmaster/master.h
@@ -78,7 +78,7 @@ struct MasterConfig
     char ip_address[MASTER_ITEM_SIZE];
     /// The port (if using TCP).
     std::uint16_t ip_port;
-    std::uint16_t rtu_baud_rate;
+    std::uint32_t rtu_baud_rate;
     std::uint8_t rtu_parity;
     std::uint16_t rtu_data_bit;
     std::uint16_t rtu_stop_bit;
